requirements.txt
selenium>=4.15.2
pytest==7.4.3
allure-pytest==2.13.2
pymysql==1.1.0
requests==2.31.0
pydantic
pytest-rerunfailures>=3.1
faker==24.9.0